Refactor notifications handling and update NotificationItem type definition
